Reproduce research ahead of dissertation submission.
Remove pipenv: this introduced more headaches than it was worth.

Running a stable debian release and installing all the software
should be more reproducible into the future, in my opinion.

The build system is a little shaky, but it appears to generate the manuscript
from a fresh pull, and they match the arxiv submission (expect the ordering of
figures in the supplemental materials).

Also removed build infrastructure for creating the paper, as all updates will be
pushed to arxiv.